+++ vmngclient/api/template_api.py
@@ -19,39 +19,15 @@ from vmngclient.api.templates.feature_template_field import FeatureTemplateField
 from vmngclient.api.templates.feature_template_payload import FeatureTemplatePayload
 from vmngclient.api.templates.models.cisco_aaa_model import CiscoAAAModel
 from vmngclient.dataclasses import Device, DeviceTemplateInfo, FeatureTemplateInfo, TemplateInfo
+from vmngclient.exceptions import AlreadyExistsError, AttachedError, TemplateNotFoundError
 from vmngclient.typed_list import DataSequence
 from vmngclient.utils.operation_status import OperationStatus
+from vmngclient.utils.template_type import TemplateType
 
 if TYPE_CHECKING:
     from vmngclient.session import vManageSession
 
 logger = logging.getLogger(__name__)
-
-
-class TemplateType(Enum):
-    CLI = "file"
-    FEATURE = "template"
-
-
-class TemplateNotFoundError(Exception):
-    """Used when a template item is not found."""
-
-    def __init__(self, template):
-        self.message = f"No such template: '{template}'"
-
-
-class AttachedError(Exception):
-    """Used when delete attached template."""
-
-    def __init__(self, template):
-        self.message = f"Template: {template} is attached to device."
-
-
-class TemplateTypeError(Exception):
-    """Used when wrong type template."""
-
-    def __init__(self, name):
-        self.message = f"Template: {name} - wrong template type."
 
 
 class DeviceTemplateFeature(Enum):
@@ -65,6 +41,27 @@ class DeviceTemplateFeature(Enum):
 class TemplatesAPI:
     def __init__(self, session: vManageSession) -> None:
         self.session = session
+
+    @overload
+    def get(self, template: Type[DeviceTemplate]) -> DataSequence[DeviceTemplateInfo]:  # type: ignore
+        ...
+
+    @overload
+    def get(self, template: Type[FeatureTemplate]) -> DataSequence[FeatureTemplateInfo]:  # type: ignore
+        ...
+
+    @overload
+    def get(self, template: Type[CLITemplate]) -> DataSequence[TemplateInfo]:  # type: ignore
+        ...
+
+    def get(self, template):
+        if template is FeatureTemplate:
+            return self._get_feature_templates()
+
+        if template in [DeviceTemplate, CLITemplate]:
+            return self._get_device_templates()
+
+        raise NotImplementedError()
 
     def _get_feature_templates(
         self, summary: bool = True, offset: Optional[int] = None, limit: Optional[int] = None
@@ -87,59 +84,17 @@ class TemplatesAPI:
         templates = self.session.get(url=endpoint, params=params)
         return templates.dataseq(DeviceTemplateInfo)
 
-    def _get_cli_templates(self) -> DataSequence[TemplateInfo]:
-        endpoint = "/dataservice/template/device"
-        templates = self.session.get(url=endpoint)
-        return templates.dataseq(TemplateInfo)
+    def attach(self, name: str, device: Device, **kwargs):
+        template_type = self.get(DeviceTemplate).filter(name=name).single_or_default().config_type
+        if template_type == TemplateType.CLI:
+            return self._attach_cli(name, device)
 
-    @overload
-    def get(self, template: Type[DeviceTemplate]) -> DataSequence[DeviceTemplateInfo]:  # type: ignore
-        ...
-
-    @overload
-    def get(self, template: Type[FeatureTemplate]) -> DataSequence[FeatureTemplateInfo]:  # type: ignore
-        ...
-
-    @overload
-    def get(self, template: Type[CLITemplate]) -> DataSequence[TemplateInfo]:  # type: ignore
-        ...
-
-    def get(self, template):
-        if template is FeatureTemplate:
-            return self._get_feature_templates()
-
-        if template is DeviceTemplate:
-            return self._get_device_templates(template)
-
-        if template is CLITemplate:
-            return self._get_cli_templates()
+        if template_type == TemplateType.FEATURE:
+            return self._attach_feature(name, device, **kwargs)
 
         raise NotImplementedError()
 
-    @overload
-    def attach(self, template: CLITemplate, name: str, device: Device) -> bool:
-        ...
-
-    @overload
-    def attach(self, template: DeviceTemplate) -> bool:
-        ...
-
-    def attach(self, template, device, name=None, **kwargs):
-        if isinstance(template, CLITemplate):
-            return self.attach_cli(template.name, device)
-
-        if template is CLITemplate and name:
-            return self.attach_cli(name, device)
-
-        if isinstance(template, DeviceTemplate):
-            return self.attach_feature(template.name, device, **kwargs)
-
-        if template is DeviceTemplate and name:
-            return self.attach_feature(name, device, **kwargs)
-
-        raise NotImplementedError()
-
-    def attach_feature(self, name: str, device: Device, **kwargs):
+    def _attach_feature(self, name: str, device: Device, **kwargs):
         """Attach Device Template created with Feature Templates.
 
         Args:
@@ -201,36 +156,6 @@ class TemplatesAPI:
         logger.warning(f"Task activity information: {task.activity}")
         return False
 
-    def edit_before_push(self, name: str, device: Device) -> bool:
-        """
-        Edits device / CLI template before pushing modified config to device(s)
-
-        Args:
-            name (str): Template name to edit.
-            device (Device): Device to attach template.
-
-        Returns:
-            bool: True if edit template is successful, otherwise - False.
-        """
-        try:
-            template_id = self.get_id(name)  # type: ignore
-            self.template_validation(template_id, device=device)
-        except TemplateNotFoundError:
-            logger.error(f"Error, Template with name {name} not found on {device}.")
-            return False
-        except HTTPError as error:
-            error_details = json.loads(error.response.text)
-            logger.error(f"Error in config: {error_details['error']['details']}.")
-            return False
-        payload = {"templateId": template_id, "deviceIds": [device.uuid], "isEdited": True, "isMasterEdited": True}
-        endpoint = "/dataservice/template/device/config/input/"
-        logger.info(f"Editing template: {name} of device: {device.hostname}.")
-        response = self.session.post(url=endpoint, json=payload).json()
-        if (response.get("data") is not None) and (response["data"][0].get("csv-status") == "complete"):
-            return True
-        logger.warning(f"Failed to edit tempate: {name} of device: {device.hostname}.")
-        return False
-
     def _attach_cli(self, name: str, device: Device, is_edited: bool = False) -> bool:
         """
 
@@ -280,27 +205,7 @@ class TemplatesAPI:
         logger.warning(f"Task activity information: {task.activity}")
         return False
 
-    @overload
-    def deattach(self, template: CLITemplate, name: str, device: Device) -> bool:
-        ...
-
-    @overload
-    def deattach(self, template: DeviceTemplate) -> bool:
-        ...
-
-    def deattach(self, template, device, name=None, **kwargs):
-        if isinstance(template, CLITemplate):
-            return self.device_to_cli(device)
-
-        if isinstance(template, DeviceTemplate):
-            raise NotImplementedError()
-
-        if template is DeviceTemplate and name:
-            raise NotImplementedError()
-
-        raise NotImplementedError()
-
-    def device_to_cli(self, device: Device) -> bool:
+    def deattach(self, device: Device) -> bool:
         """
 
         Args:
@@ -409,6 +314,50 @@ class TemplatesAPI:
             raise AttachedError(template.name)
         return True
 
+    @overload
+    def create(self, template: FeatureTemplate) -> str:
+        ...
+
+    @overload
+    def create(self, template: DeviceTemplate) -> str:
+        ...
+
+    @overload
+    def create(self, template: CLITemplate) -> str:
+        ...
+
+    def create(self, template, debug: bool = False):
+        if isinstance(template, list):
+            return [self.create(t) for t in template]
+
+        template_id: Optional[str] = None  # type: ignore
+        template_type = None
+
+        exists = self.get(type(template)).filter(name=template.name)
+        if exists:
+            raise AlreadyExistsError(f"Template with name [{template.name}] already exists.")
+
+        if isinstance(template, FeatureTemplate):
+            if self.is_created_by_generator(template):
+                template_id = self.create_by_generator(template, debug)
+            else:
+                template_id = self._create_feature_template(template)
+            template_type = FeatureTemplate.__name__
+
+        if isinstance(template, DeviceTemplate):
+            template_id = self._create_device_template(template)
+            template_type = DeviceTemplate.__name__
+
+        if isinstance(template, CLITemplate):
+            template_id = self._create_cli_template(template)
+            template_type = CLITemplate.__name__
+
+        if not template_id:
+            raise NotImplementedError()
+
+        logger.info(f"Template {template.name} ({template_type}) was created successfully ({template_id}).")
+        return template_id
+
     def _create_feature_template(self, template: FeatureTemplate) -> str:
         payload = template.generate_payload(self.session)
         response = self.session.post("/dataservice/template/feature", json=json.loads(payload))
@@ -460,50 +409,6 @@ class TemplatesAPI:
         response = self.session.post(endpoint, json=payload)
 
         return response.text
-
-    @overload
-    def create(self, template: FeatureTemplate) -> str:
-        ...
-
-    @overload
-    def create(self, template: DeviceTemplate) -> str:
-        ...
-
-    @overload
-    def create(self, template: CLITemplate) -> str:
-        ...
-
-    def create(self, template, debug: bool = False):
-        if isinstance(template, list):
-            return [self.create(t) for t in template]
-
-        template_id: Optional[str] = None  # type: ignore
-        template_type = None
-
-        # exists = self.get(type(template)).filter(name=template.name)
-        # if exists:
-        #     raise AlreadyExistsError(f"Template [{template.name}] already exists.")
-
-        if isinstance(template, FeatureTemplate):
-            if self.is_created_by_generator(template):
-                template_id = self.create_by_generator(template, debug)
-            else:
-                template_id = self._create_feature_template(template)
-            template_type = FeatureTemplate.__name__
-
-        if isinstance(template, DeviceTemplate):
-            template_id = self._create_device_template(template)
-            template_type = DeviceTemplate.__name__
-
-        if isinstance(template, CLITemplate):
-            template_id = self._create_cli_template(template)
-            template_type = CLITemplate.__name__
-
-        if not template_id:
-            raise NotImplementedError()
-
-        logger.info(f"Template {template.name} ({template_type}) was created successfully ({template_id}).")
-        return template_id
 
     def is_created_by_generator(self, template: FeatureTemplate) -> bool:
         """Checks if template is created by generator
@@ -593,3 +498,33 @@ class TemplatesAPI:
         endpoint = "/dataservice/template/device/config/config/"
         response = self.session.post(url=endpoint, json=payload)
         return response.text
+
+    def edit_before_push(self, name: str, device: Device) -> bool:
+        """
+        Edits device / CLI template before pushing modified config to device(s)
+
+        Args:
+            name (str): Template name to edit.
+            device (Device): Device to attach template.
+
+        Returns:
+            bool: True if edit template is successful, otherwise - False.
+        """
+        try:
+            template_id = self.get(CLITemplate).filter(name=name).single_or_default().id
+            self.template_validation(template_id, device=device)
+        except TemplateNotFoundError:
+            logger.error(f"Error, Template with name {name} not found on {device}.")
+            return False
+        except HTTPError as error:
+            error_details = json.loads(error.response.text)
+            logger.error(f"Error in config: {error_details['error']['details']}.")
+            return False
+        payload = {"templateId": template_id, "deviceIds": [device.uuid], "isEdited": True, "isMasterEdited": True}
+        endpoint = "/dataservice/template/device/config/input/"
+        logger.info(f"Editing template: {name} of device: {device.hostname}.")
+        response = self.session.post(url=endpoint, json=payload).json()
+        if (response.get("data") is not None) and (response["data"][0].get("csv-status") == "complete"):
+            return True
+        logger.warning(f"Failed to edit tempate: {name} of device: {device.hostname}.")
+        return False
