+++ vmngclient/utils/template_type.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+
+class TemplateType(Enum):
+    CLI = "file"
+    FEATURE = "template"
